requirements.txt
hydra-core==1.3.2
matplotlib==3.8.2
numpy==1.26.3
pandas==2.1.4
Pillow==10.2.0
scikit_learn==1.3.2
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
wandb==0.16.1
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
dvc==3.38.1
dvc[gs]
pytest==7.4.4
